requirements.txt
# Editable Git install with no remote (aki-prediction==0.1.0)
-e /home/rahul/desktop/mhedas/1st_year/q1/electronic_health_records/assessments/a03_aki_prediction
asttokens==3.0.0
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
executing==2.2.0
fonttools==4.58.5
imbalanced-learn==0.13.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.1
kiwisolver==1.4.8
matplotlib==3.10.3
matplotlib-inline==0.1.7
numpy==2.3.1
packaging==25.0
pandas==2.3.0
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.3.0
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.16.0
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.3
stack-data==0.6.3
statsmodels==0.14.4
threadpoolctl==3.6.0
traitlets==5.14.3
tzdata==2025.2
wcwidth==0.2.13
